+++ crossover-server-2/src/main/java/org/example/crossoverserver2/planeletter/controller/BoardController.java
@@ -2,7 +2,7 @@ package org.example.crossoverserver2.planeletter.controller;
 
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
-import org.example.crossoverserver2.planeletter.dto.request.WriteBoardDto;
+import org.example.crossoverserver2.planeletter.dto.request.board.WriteBoardDto;
 import org.example.crossoverserver2.planeletter.dto.response.board.BoardListResponseData;
 import org.example.crossoverserver2.planeletter.dto.response.board.BoardResponseData;
 import org.example.crossoverserver2.planeletter.dto.ResponseDto;
