+++ crossover-server-2/src/main/java/org/example/crossoverserver2/planeletter/dto/request/board/WriteBoardDto.java
@@ -1,4 +1,4 @@
-package org.example.crossoverserver2.planeletter.dto.request;
+package org.example.crossoverserver2.planeletter.dto.request.board;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
