+++ crossover-server-2/src/main/java/org/example/crossoverserver2/planeletter/model/Board.java
@@ -13,21 +13,17 @@ import java.util.List;
 public class Board extends BaseEntity{
 
     @Setter
-    @Column(length = 20, nullable = false)// 길이는 100자 이하이고, 비어있을 수 없다.
+    @Column(length = 20, nullable = false)// 길이는 20자 이하이고, 비어있을 수 없다.
     private String title;
 
     @Setter
-    @Column(length = 140, nullable = false)// 길이는 2000자 이하이고, 비어있을 수 없다.
+    @Column(length = 140, nullable = false)// 길이는 140자 이하이고, 비어있을 수 없다.
     private String content;
-
 
     @ManyToOne(optional = false, fetch = FetchType.EAGER)
     private User user;
 
     @OneToMany(mappedBy = "board", cascade = CascadeType.ALL, orphanRemoval = true, fetch = FetchType.EAGER)
     private List<Comment> comments;
-
-
-
 
 }
